requirements.txt
pyautogui
pytesseract
pillow
pyqt5
keyboard
pydantic